+++ src/gui/EmulatorUI.java
@@ -22,7 +22,7 @@ public abstract class EmulatorUI  { // if there is no non-abstract methods, turn
 	
 	public abstract void reset();
 	
-	public abstract void updateMemory(int address, byte data);
+	public abstract void updateMemory(int address, int data);
 	
 	public abstract void setSource(String source);
